chore/deps: upgrade libsodium version to 1.0.16

+++ src/crypto/auth/hmacsha512.rs
@@ -8,8 +8,8 @@ use ffi::{crypto_auth_hmacsha512, crypto_auth_hmacsha512_BYTES, crypto_auth_hmac
 auth_module!(
     crypto_auth_hmacsha512,
     crypto_auth_hmacsha512_verify,
-    crypto_auth_hmacsha512_KEYBYTES,
-    crypto_auth_hmacsha512_BYTES
+    crypto_auth_hmacsha512_KEYBYTES as usize,
+    crypto_auth_hmacsha512_BYTES as usize
 );
 
 auth_state!(
