+++ src/crypto/stream/chacha20.rs
@@ -6,8 +6,8 @@ use ffi::{crypto_stream_chacha20, crypto_stream_chacha20_KEYBYTES,
 stream_module!(
     crypto_stream_chacha20,
     crypto_stream_chacha20_xor,
-    crypto_stream_chacha20_KEYBYTES,
-    crypto_stream_chacha20_NONCEBYTES
+    crypto_stream_chacha20_KEYBYTES as usize,
+    crypto_stream_chacha20_NONCEBYTES as usize
 );
 
 #[cfg(test)]
