+++ src/crypto/stream/xsalsa20.rs
@@ -9,8 +9,8 @@ use ffi::{crypto_stream_xsalsa20, crypto_stream_xsalsa20_KEYBYTES,
 stream_module!(
     crypto_stream_xsalsa20,
     crypto_stream_xsalsa20_xor,
-    crypto_stream_xsalsa20_KEYBYTES,
-    crypto_stream_xsalsa20_NONCEBYTES
+    crypto_stream_xsalsa20_KEYBYTES as usize,
+    crypto_stream_xsalsa20_NONCEBYTES as usize
 );
 
 #[cfg(test)]
